5. SAP Generative AI Hub and HANA Cloud Vector Engine/5.1 Q&A RAG/code/ai-core-rag-pipeline/requirements.txt
fastapi
uvicorn
langchain
openai
docarray
panel
jupyter_bokeh
redlines
tiktoken
wikipedia
DateTime
presidio_anonymizer
hana-ml
hdbcli
presidio_analyzer
generative-ai-hub-sdk[all]
python-dotenv